delete: 第一阶段
