+++ dice/platform_adapter_lagrange_helper_others.go
